fix gRPC Product Hub tests

See merge request qa/automag!891

+++ web-tests/grpc/src/test/java/ru/instamart/test/content/product_hub/ProductHubBackTest.java
@@ -448,7 +448,7 @@ public class ProductHubBackTest extends GrpcBase {
                         .addOfferData(Data.newBuilder()
                                 .setKey("popularity")
                                 .addValues("20")
-                                .setType(ValueType.INTEGER)
+                                .setType(ValueType.DOUBLE)
                                 .build())
                         .setStatus(Status.ENABLE)
                         .build())
